refactor: :recycle: beerListFilterChips atom 제거 및 초기화 로직 훅으로 분리
