+++ src/pages/beers/components/filter/BeerListFilterAndSorter/BeerListFilterAndSorter.tsx
@@ -1,25 +1,21 @@
 import styled from '@emotion/styled';
-import _ from 'lodash';
-import { useEffect } from 'react';
 import { useRecoilState } from 'recoil';
 
-import { useGetBeerTypes } from '@/apis/beerTypes/getBeerTypes';
-import { useGetCountries } from '@/apis/countries/getCountries';
 import { HEADER_HEIGHT } from '@/components/Header';
 import { useModal } from '@/hooks';
 import {
   $beerListFilter,
-  $beerListFilterChips,
   $beerListOrder,
   beerListOrderTextAlias,
 } from '@/pages/beers/recoil/atoms';
-import { BeerListFilter, BeerListFilterChip, IBeerType, ICountry } from '@/types';
+import { BeerListFilter, BeerListFilterChip } from '@/types';
 
 import BeerListFilterBottomSheet from '../BeerListFilterBottomSheet';
 import BeerListFilterChipList from '../BeerListFilterChipList';
 import BeerListSortBottomSheet from '../BeerListSortBottomSheet';
 
 import FilterButton from './FilterButton';
+import { useFilterChipsState } from './hooks/useFilterChipsState';
 import SortButton from './SortButton';
 
 interface BeerListFilterAndSorterProps {
@@ -27,23 +23,14 @@ interface BeerListFilterAndSorterProps {
   totalCount?: number;
 }
 
-// TODO: filter, order -> QueryParams 연동
+// TODO: filter, order, keyword (?) -> QueryParams 연동
 const BeerListFilterAndSorter = ({ resultCount, totalCount }: BeerListFilterAndSorterProps) => {
-  const { data: beerTypes = [] } = useGetBeerTypes();
-  const { data: countries = [] } = useGetCountries();
-
   const [filter, setFilter] = useRecoilState($beerListFilter);
-  const [filterChips, setFilterChips] = useRecoilState($beerListFilterChips);
+  const [filterChips, setFilterChips] = useFilterChipsState(filter);
   const [order, setOrder] = useRecoilState($beerListOrder);
 
   const [isFilterBottomSheetOpen, openFilterBottomSheet, closeFilterBottomSheet] = useModal(false);
   const [isSortBottomSheetOpen, openSortBottomSheet, closeSortBottomSheet] = useModal(false);
-
-  useEffect(() => {
-    if (!_.isEmpty(filter) || !beerTypes.length || !countries.length) return;
-
-    setFilterChips(initFilterChips({ filter, beerTypes, countries }));
-  }, [beerTypes, countries, filter, setFilterChips]);
 
   const removeFilterChip = (chip: BeerListFilterChip) => {
     setFilter({
@@ -108,36 +95,3 @@ const StyledWrapper = styled.div`
     }
   }
 `;
-
-const getBeerTypeNameKor = (beerTypes: IBeerType[], typeId: IBeerType['id']) => {
-  return beerTypes.find((beerType) => beerType.id === typeId)?.korName || '';
-};
-
-const getCountryNameKor = (countries: ICountry[], countryId: ICountry['id']) => {
-  return countries.find((country) => country.id === countryId)?.korName || '';
-};
-
-const initFilterChips = ({
-  filter,
-  beerTypes,
-  countries,
-}: {
-  filter: BeerListFilter;
-  beerTypes: IBeerType[];
-  countries: ICountry[];
-}): BeerListFilterChip[] => {
-  const { type = [], country = [] } = filter;
-
-  return [
-    ...type.map((typeId) => ({
-      id: typeId,
-      text: getBeerTypeNameKor(beerTypes, typeId),
-      filterKey: 'type' as const,
-    })),
-    ...country.map((countryId) => ({
-      id: countryId,
-      text: getCountryNameKor(countries, countryId),
-      filterKey: 'country' as const,
-    })),
-  ];
-};
