+++ src/pages/beers/recoil/atoms/index.ts
@@ -1,5 +1,4 @@
 export { BEER_LIST_FILTER_ATOM_KEY, $beerListFilter } from './beerListFilter';
-export { BEER_LIST_FILTER_CHIPS_ATOM_KEY, $beerListFilterChips } from './beerListFilterChips';
 export {
   BEER_LIST_ORDER_ATOM_KEY,
   beerListOrderTextAlias,
